Updated material and tried to get no skybox working

+++ Lamp/src/Platform/Vulkan/VulkanRenderPipeline.h
@@ -29,7 +29,7 @@ namespace Lamp
 		void SetPushConstantData(Ref<CommandBuffer> commandBuffer, uint32_t index, const void* data);
 		const RenderPipelineSpecification& GetSpecification() const override { return m_specification; }
 
-		inline const uint32_t GetDescriptorSetCount() const { return m_descriptorSets.at(0).size(); }
+		inline const uint32_t GetDescriptorSetCount() const { return (uint32_t)m_descriptorSets.at(0).size(); }
 
 	private:
 		void CreateDescriptorSets();
